+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 development:
   - add --period to "synccommittee members", can be "current", "next"
+  - add "validator expectation"
 
 1.14.0:
   - add "chain verify signedcontributionandproof"